+++ packages/tswebextension/src/lib/common/tabs/tabs-api.ts
@@ -403,6 +403,38 @@ export abstract class TabsApiCommon<F extends FrameCommon, T extends TabContextC
     }
 
     /**
+     * Sets a current timestamp as `assistantInitTimestamp` of the tab context.
+     *
+     * Needed to determine later if a newly created frame is an assistant frame.
+     *
+     * @param tabId Tab id.
+     */
+    public setAssistantInitTimestamp(tabId: number): void {
+        const tabContext = this.context.get(tabId);
+
+        if (!tabContext) {
+            return;
+        }
+
+        tabContext.assistantInitTimestamp = Date.now();
+    }
+
+    /**
+     * Resets tab context's `assistantInitTimestamp` to null.
+     *
+     * @param tabId Tab id.
+     */
+    public resetAssistantInitTimestamp(tabId: number): void {
+        const tabContext = this.context.get(tabId);
+
+        if (!tabContext) {
+            return;
+        }
+
+        tabContext.assistantInitTimestamp = null;
+    }
+
+    /**
      * Sets frame context.
      *
      * @param tabId Tab ID.